Update Particle3D.java
So for some reason some of the changes we made yesterday seem to have vanished, so I tried to recreate them.
Eclipse is complaining about vectorSubtraction and I can't work out why. Seems to be the only Vector3D method it doesn't like though, so I don't think the file is saved in the wrong place or anything.
Also I'm not sure what you did about addScaledVector?